% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkdata.R
\name{linkdata}
\alias{linkdata}
\title{linkdata}
\usage{
linkdata(Dsmurf, Dfish, Dswath, Meta, M, pool.sides = TRUE, interp = FALSE)
}
\arguments{
\item{Dswath}{kelp swath dataframe}

\item{Meta}{Metadata associated with site IDs}

\item{M}{Natural mortality rate (for calculating discounted recruitment index)}

\item{pool.sides}{Pool both 'sides' of the study site (default TRUE)}

\item{interp}{Interpolate missing data years (default FALSE)}

\item{dSmurf}{SMURF dataframe}

\item{dFish}{subtidal fish dataframe}
}
\value{
Combined dataframe. Also returns digits counting through the loop over sites and years, to show progress
}
\description{
Function for combining SMURF, subtidal fish, and kelp swath dataframes into a single dataframe
}
