% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBTLdata.R
\name{SBTLdata}
\alias{SBTLdata}
\title{SBTLdata}
\usage{
SBTLdata(Species, MinLen, pool.sides = TRUE, Names = FALSE)
}
\value{
Dataframe
}
\description{
Read in PISCO subtidal visual transect data and format into dataframe
}
